feat: Splitting code into core components
