+++ build.gradle
@@ -15,6 +15,12 @@ plugins {
     id 'net.researchgate.release' version '2.8.1'
 }
 
+allprojects {
+    repositories {
+        mavenLocal()
+    }
+}
+
 def branch = System.getenv("GITHUB_REF")
 if(branch != null) {
     branch = branch.replace('refs/heads/', '')
@@ -33,7 +39,6 @@ release {
     git {
         requireBranch = ''
     }
-
 
     // Disable tasks because something we have is causing -x to be ignored
     createReleaseTag.enabled = false
