+++ core/src/main/java/com/sekwah/reskin/core/client/SkinLoadJob.java
@@ -1,4 +1,4 @@
-package com.sekwah.reskin.client;
+package com.sekwah.reskin.core.client;
 
 import java.util.UUID;
