+++ fabric/build.gradle
@@ -25,6 +25,9 @@ loom {
 
 sourceSets {
 	main {
+		java {
+			srcDir "../core/src/main/java"
+		}
 		resources {
 			srcDirs "../core/src/main/resources"
 		}
@@ -55,12 +58,21 @@ processResources {
 	}
 }
 
+// Stops trying to resolve forge dependencies while doing fabric
+configurations {
+	all {
+		exclude group: "net.minecraftforge"
+	}
+}
+
 dependencies {
 	//to change the versions see the gradle.properties file
 	minecraft "com.mojang:minecraft:1.17.1"
 	mappings minecraft.officialMojangMappings()
 	modImplementation "net.fabricmc:fabric-loader:0.11.6"
 	modImplementation "net.fabricmc.fabric-api:fabric-api:0.37.1+1.17"
+
+	implementation project(":core")
 
 	modImplementation "com.terraformersmc:modmenu:2.0.5"
