--- forge/src/main/java/com/sekwah/reskin/client/ClientSkinData.java
@@ -1,14 +1,0 @@
-package com.sekwah.reskin.client;
-
-import net.minecraft.resources.ResourceLocation;
-
-public class ClientSkinData {
-
-    public final ResourceLocation resourceLocation;
-    public final String modelType;
-
-    public ClientSkinData(ResourceLocation resourceLocation, String modelType) {
-        this.resourceLocation = resourceLocation;
-        this.modelType = modelType;
-    }
-}
