+++ forge/src/main/java/com/sekwah/reskin/client/ClientSkinManager.java
@@ -3,6 +3,9 @@ package com.sekwah.reskin.client;
 import com.google.common.collect.Maps;
 import com.mojang.authlib.minecraft.MinecraftProfileTexture;
 import com.sekwah.reskin.ReSkin;
+import com.sekwah.reskin.core.client.ClientSkinData;
+import com.sekwah.reskin.core.client.HDDownloadingTexture;
+import com.sekwah.reskin.core.client.SkinLoadJob;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.player.AbstractClientPlayer;
 import net.minecraft.client.renderer.texture.TextureManager;
